[Minjin.Kim] [CON-68] 프로젝트 공고 생성 API 연결 - 프로젝트 공고 생성 API 연결 - 리프래시 토큰 갱신 로직 변경

+++ domain/src/main/java/com/connectcrew/domain/usecase/project/entity/ProjectFeedLikeInfoEntity.kt
@@ -3,7 +3,7 @@ package com.connectcrew.domain.usecase.project.entity
 import com.connectcrew.domain.usecase.Entity
 
 data class ProjectFeedLikeInfoEntity(
-    val projectId: Int,
+    val projectId: Long,
     val isLike: Boolean,
     val likeCount: Int
 ) : Entity
